+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.26.5] TBD
+## [0.26.5] 2024-04-28
 
 - **Breaking change** - changed the api of `updateMany` to receive a `set` option, instead of second parameter for the set
 - Fixed primary key was not created for entities that had more than one id column using knex or postgres